Rename onboarding and packages
