+++ samples/app/src/main/kotlin/com/bumble/appyx/app/node/root/RootNode.kt
@@ -1,7 +1,6 @@
 package com.bumble.appyx.app.node.root
 
 import android.content.Context
-import android.content.Context.MODE_PRIVATE
 import android.os.Parcelable
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.animation.core.tween
@@ -17,11 +16,10 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.ExperimentalUnitApi
-import androidx.core.content.edit
 import com.bumble.appyx.R
 import com.bumble.appyx.app.composable.ScreenCenteredContent
 import com.bumble.appyx.app.node.helper.screenNode
-import com.bumble.appyx.app.node.onboarding.OnboardingContainerNode
+import com.bumble.appyx.app.node.slideshow.WhatsAppyxSlideShow
 import com.bumble.appyx.app.node.root.RootNode.NavTarget
 import com.bumble.appyx.app.node.samples.SamplesContainerNode
 import com.bumble.appyx.core.composable.Children
@@ -54,7 +52,7 @@ class RootNode(
         object Splash : NavTarget()
 
         @Parcelize
-        object Onboarding : NavTarget()
+        object WhatsAppyx : NavTarget()
 
         @Parcelize
         object Samples : NavTarget()
@@ -63,13 +61,13 @@ class RootNode(
     override fun resolve(navTarget: NavTarget, buildContext: BuildContext): Node =
         when (navTarget) {
             NavTarget.Splash -> screenNode(buildContext) { Splash() }
-            NavTarget.Onboarding -> OnboardingContainerNode(buildContext)
+            NavTarget.WhatsAppyx -> WhatsAppyxSlideShow(buildContext)
             NavTarget.Samples -> SamplesContainerNode(buildContext)
         }
 
     override fun onChildFinished(child: Node) {
         when (child) {
-            is OnboardingContainerNode -> backStack.newRoot(NavTarget.Samples)
+            is WhatsAppyxSlideShow -> backStack.newRoot(NavTarget.Samples)
             else -> super.onChildFinished(child)
         }
     }
