+++ samples/app/src/main/kotlin/com/bumble/appyx/app/node/slideshow/slide/modeldriven/ComposableNavigation.kt
@@ -1,4 +1,4 @@
-package com.bumble.appyx.app.node.onboarding.screen.modeldriven
+package com.bumble.appyx.app.node.slideshow.slide.modeldriven
 
 import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
